% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ac1mon}
\alias{ac1mon}
\title{Estimation of lag-1 month-to-month correlation coefficients.}
\usage{
ac1mon(data)
}
\arguments{
\item{data}{}
}
\value{
a vector
}
\description{
Estimation of lag-1 month-to-month correlation coefficients.
}
