% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deseason_trend}
\alias{deseason_trend}
\title{Deseason and detrend}
\usage{
deseason_trend(x, win.len = 12, do.plot = FALSE)
}
\arguments{
\item{x}{a vector}

\item{win.len}{window length}

\item{do.plot}{a logic value}
}
\value{
a vector
}
\description{
Deseason and detrend
}
