% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_annual_to_daily.R
\name{knn_annual_to_daily}
\alias{knn_annual_to_daily}
\title{Annual to daily disaggregation}
\usage{
knn_annual_to_daily(Q1, Q2, K)
}
\arguments{
\item{Q1}{An annual dataframe (year, Qa).}

\item{Q2}{A daily dataframe (year, month, day, Qd).}

\item{K}{Number of nearest neighbours.}
}
\value{
A data frame (year, nn, month, day, Qd) (nn for the sampled one of the k nearest neighbours).
}
\description{
Disaggregate Q1 by sampling from Q2 using (modified) k-nearest-neighbour resampling (Nowalk et al, 2010).
}
\references{
Nowak, K., J. Prairie, B. Rajagopalan, and U. Lall (2010), A nonparametric stochastic approach for multisite disaggregation of annual to daily streamflow, Water Resour. Res., 46, W08529, doi:10.102/2009WR008530.
}
