% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric}
\alias{metric}
\title{Typical metrics for regression}
\usage{
metric(sim, obs, use = "pairwise", method = "pearson")
}
\arguments{
\item{sim}{simulated values}

\item{obs}{observed values}

\item{use}{everything or pairwise}

\item{method}{a character string indicating which correlation coefficient}
}
\value{
a vector of metric
}
\description{
Typical metrics for regression
}
\examples{
NULL
}
