% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{NSE}
\alias{NSE}
\title{Nash-Sutcliffe model efficiency}
\usage{
NSE(sim, obs, j = 1)
}
\arguments{
\item{sim}{simulated values}

\item{obs}{observed values}

\item{j}{the exponent to be used in the computation of the modified Nash-Sutcliffe effciency. The default value is j=1}
}
\value{
a vector of metric
}
\description{
Nash-Sutcliffe model efficiency
}
\examples{
NULL
}
