% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_mat_lag}
\alias{pred_mat_lag}
\title{Predictor format}
\usage{
pred_mat_lag(pred.mat, lag = 3, mv = 12)
}
\arguments{
\item{pred.mat}{a matrix}

\item{lag}{max no of lag for predictors}

\item{mv}{moving window}
}
\value{
a matrix
}
\description{
Predictor format
}
