\name{qgen-package}
\alias{qgen-package}
\alias{qgen}
\docType{package}
\title{
   Quantitative Genetics in R
}
\description{
Functions to estimate parameters and compute statistics and their corresponding confidence intervals (e.g. heritability, genetic correlations...). All functions are written for sib analysis with up to 2 levels of nesting. Parametric simulations allow the calculation of bootstrap confidence intervals and the evaluation of dependence on different assumptions.
}
\details{
Further information is available in the following
\link[utils:vignette]{vignette}:
\tabular{ll}{
  \code{overview} \tab An overview with examples (source, pdf)\cr
}

For a complete list of functions with individual help pages,
use \code{library(help="qgen")}.
}
\author{

Thomas Fabbro \email{thomas.fabbro@unibas.ch}

}

\keyword{ package }