% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time2season}
\alias{time2season}
\title{Convert date to season format}
\usage{
time2season(x, out.fmt = "months", type = "default")
}
\arguments{
\item{x}{date}

\item{out.fmt}{output format}

\item{type}{season type}
}
\value{
a season
}
\description{
Convert date to season format
}
